feat!: handle identity CIDs (#10)
Identity CIDs use the `identity` multihash code `0x0`, to specify that
the multihash digest is the data itself rather than a hash.

Identity CIDs can be encountered anywhere you could find a regular `cid,
bytes` pair:
- as an explicit entry in the CAR with its's bytes as a 0 length
Uint8Array
- as a link target in the decoded bytes of the of a non-identity CID
- as a link target in an other identity CID (!)

I've chosen to define a given set of block to index as being a
**Complete** DAG where a block links to Identity CIDs, even if those
CIDs are not also part of the set of blocks to index, as we have all the
data we need to complete the DAG. There is a tangentially related note
about this concept in the indexing section of the CARv2 spec

> Indexes should not include identity hash CIDs unless the fully-indexed
characteristic is set. It is assumed that any use of a CARv2 as a
blockstore will return identity CID data immediately by extracting it
from the CID, therefore there should be no need to provide indexing for
such entries. However, when fully-indexed characteristic is set, the
blockstore should persist blocks with identity CID into the CARv2 data
payload and index them.
– https://ipld.io/specs/transport/car/carv2/#index-format

_note_ linkdex as currently implemented is not a spec compliant CARv2
index, and it is not yet aiming to be. This note is just useful context
to see another example of handling identity CIDs.

I've also updated the linkdex report output to more accurately reflect
what is being counted:
- `blocksIndexed` is the count of the number of blocks that the
LinkIndexer was asked to index... This does not include identity CIDs
that only appear as link targets. It does in include identity CIDs that
appear as explicit block entires. If the caller asks us to index it,
then it's counted as a block.
- `uniqueCids` is the count of **all** unique CIDs encountered,
including both explicitly indexed identity CIDs and implicitly indexed
Identity CIDs found as link targets.
- `undecodable` is the count of undecodable items, including blocks and
identity CIDs... DAG completeness is unknowable if we encounter an
undecodable item of either sort.

License: MIT
Signed-off-by: Oli Evans <oli@protocol.ai>

Signed-off-by: Oli Evans <oli@protocol.ai>
Co-authored-by: Alan Shaw <alan.shaw@protocol.ai>