feat!: support index keys as CIDs (#24)
This PR adds 2 dependencies `lnmap` and `lnset` which are `Map`/`Set`
implementations that allow keys to be CIDs. It means that
`linkIndexer.idx` is a bit more useful. For my particular usecase I want
to use the linkdex information to create content claims, which all need
`Link`s. The alternative of iterating over `linkIndexer.idx` and parsing
each CIDs would be a bit awkward and a perf hit.

The tests seem to run in roughly the same time and using these deps also
means that in the unlikely event that two indexed blocks link to the
same CID but using different IPLD codecs it'll not be indexed as 2
blocks :trollface: